New: Don't block imports when release was matched by ID if they were grabbed via interactive search
Closes #5043
