+++ src/NzbDrone.Core/History/HistoryService.cs
@@ -168,6 +168,7 @@ namespace NzbDrone.Core.History
                 history.Data.Add("Protocol", ((int)message.Episode.Release.DownloadProtocol).ToString());
                 history.Data.Add("CustomFormatScore", message.Episode.CustomFormatScore.ToString());
                 history.Data.Add("SeriesMatchType", message.Episode.SeriesMatchType.ToString());
+                history.Data.Add("ReleaseSource", message.Episode.ReleaseSource.ToString());
 
                 if (!message.Episode.ParsedEpisodeInfo.ReleaseHash.IsNullOrWhiteSpace())
                 {
